Correccion de errores

+++ odoo/addons/jcd_Fruteria/models/Fruta.py
@@ -9,7 +9,7 @@ class Fruta(models.Model):
     fruta_id = fields.Char('ID',required=True)
     name = fields.Char('Nombre',required=True)
     precio = fields.Float('Precio €/KG')
-    cantidad = fields.Double('Cantidad(KG)')
+    cantidad = fields.Float('Cantidad(KG)')
     estado = fields.Selection([
     ('1', 'Bueno'),
     ('2', 'Regular'),
